requirements.txt
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dnspython==1.16.0
eventlet==0.30.2
faunadb==4.1.0
Flask==2.0.1
Flask-SocketIO==5.0.3
future==0.18.2
greenlet==1.1.0
gunicorn==20.1.0
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.10
importlib-metadata==4.0.1
iso8601==0.1.14
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.17.1
python-engineio==4.2.0
python-socketio==5.3.0
pytz==2021.1
requests==2.25.1
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==2.0.1
zipp==3.4.1
stripe=4.0.2